Update Boots to Smee, put all services behind a single IP:
Update to use Smee. The CLI flags changed so the bulk
of the work was to rework the values.yaml for Smee
to accommodate.

Second change was to get all service behind a single
load balancer IP. All stack services except DHCP run
through nginx. A DHCP relay side car is run in the pod
with nginx to handle broadcast and unicast DHCP.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>